ts-template more specialized to ts and double types to avoid inference with other frameworks using operators
